+++ iocore/net/quic/QUICPacketRetransmitter.cc
@@ -61,7 +61,7 @@ QUICPacketRetransmitter::will_generate_frame()
 }
 
 QUICFrameUPtr
-QUICPacketRetransmitter::generate_frame(uint16_t connection_credit, uint16_t maximum_frame_size)
+QUICPacketRetransmitter::generate_frame(uint64_t connection_credit, uint16_t maximum_frame_size)
 {
   QUICFrameUPtr frame = QUICFrameFactory::create_null_frame();
   if (!this->_retransmission_frames.empty()) {
