+++ iocore/net/quic/QUICPacketRetransmitter.h
@@ -35,7 +35,7 @@ public:
 
   // QUICFrameGenerator
   bool will_generate_frame() override;
-  QUICFrameUPtr generate_frame(uint16_t connection_credit, uint16_t maximum_frame_size) override;
+  QUICFrameUPtr generate_frame(uint64_t connection_credit, uint16_t maximum_frame_size) override;
 
 private:
   std::queue<QUICFrameUPtr> _retransmission_frames;
