+++ iocore/net/quic/QUICStreamManager.h
@@ -65,7 +65,7 @@ public:
 
   // QUICFrameGenerator
   bool will_generate_frame() override;
-  QUICFrameUPtr generate_frame(uint16_t connection_credit, uint16_t maximum_frame_size) override;
+  QUICFrameUPtr generate_frame(uint64_t connection_credit, uint16_t maximum_frame_size) override;
 
 private:
   QUICStream *_find_stream(QUICStreamId id);
